Speed composer dependency parsing up
