+++ components-rs/Cargo.toml
@@ -17,7 +17,9 @@ datadog-sidecar-ffi = { path = "../libdatadog/sidecar-ffi" }
 spawn_worker = { path = "../libdatadog/spawn_worker" }
 anyhow = { version = "1.0" }
 const-str = "0.5.6"
-json = "0.12.4"
+serde = "1.0.196"
+simd-json = "0.13.8"
+serde_with = "3.6.0"
 lazy_static = "1.4"
 log = "0.4.20"
 env_logger = "0.10.1"
